Widen RSNXE capability bitmask from u32 to u64 in AP mode
WLAN_RSNX_CAPAB_UNAUTH_EPPKE is defined at bit position >= 32, which
overflows a u32 bitmask. Widen the capab local variables and the
rsnxe_capab() return type to u64 in hostapd_eid_rsnxe(),
wpa_write_rsnxe(), wpa_write_rsnxe_override(), and rsnxe_capab() so that
high-numbered RSNXE capability bits can be represented correctly.

+++ src/ap/ieee802_11_shared.c
@@ -1115,7 +1115,7 @@ u8 * hostapd_eid_rsnxe(struct hostapd_data *hapd, u8 *eid, size_t len)
 {
 	u8 *pos = eid;
 	bool sae_pk = false;
-	u32 capab = 0, tmp;
+	u64 capab = 0, tmp;
 	size_t flen;
 
 	if (!(hapd->conf->wpa & WPA_PROTO_RSN))
